refactor: split main.cpp into cli, jit_runner, and test_runner modules
Split the 1018-line main.cpp into separate modules for better
separation of concerns:

- cli.cpp: CLI flag parsing, help text, entry point resolution
- jit_runner.cpp: JIT compilation pipeline, CoverageState management
- test_runner.cpp: test file discovery, sequential/parallel execution

Also consolidate small codegen files:
- codegen_print.cpp (25 lines) merged into codegen_call_io.cpp
- codegen_call_path.cpp (16 lines) merged into codegen_call_io.cpp

Global state eliminated: g_skip_global_lib, g_trace_enabled,
g_trace_out, and g_coverage_* replaced with function parameters
and CoverageState struct.

Closes #677
