+++ src/main.cpp
@@ -1,791 +1,30 @@
-#include "ry/lexer.hpp"
-#include "ry/parser.hpp"
-#include "ry/module_loader.hpp"
-#include "ry/codegen.hpp"
-#include "ry/source_manager.hpp"
-#include "ry/diagnostic.hpp"
-#include "ry/jit.hpp"
-#include "ry/test_runtime.hpp"
+#include "ry/cli.hpp"
+#include "ry/jit_runner.hpp"
+#include "ry/test_runner.hpp"
 #include "ry/project_config.hpp"
 #include "ry/formatter.hpp"
 #include "ry/self_update.hpp"
 #include "ry/args_runtime.hpp"
-#include "ry/paths.hpp"
 #include "ry/file_watcher.hpp"
-#include "ry/coverage_runtime.hpp"
 #include "ry/trace.hpp"
-#include "ry/dotenv.hpp"
-#include <algorithm>
-#include <cerrno>
-#include <chrono>
-#ifdef __APPLE__
-#include <crt_externs.h>
-#define RY_ENVIRON (*_NSGetEnviron())
-#else
-extern "C" { extern char **environ; }
-#define RY_ENVIRON environ
-#endif
-#include <cstdlib>
-#include <iostream>
-#include <csignal>
-#include <cstdio>
+#include "ry/diagnostic.hpp"
 #include <cstring>
-#include <deque>
 #include <filesystem>
-#include <fstream>
-#include <mutex>
-#include <spawn.h>
-#include <sys/wait.h>
-#include <thread>
-#include <unistd.h>
-#include <unordered_set>
-#include <vector>
-#include <llvm/ExecutionEngine/Orc/LLJIT.h>
-#include <llvm/ExecutionEngine/Orc/ExecutorProcessControl.h>
-#include <llvm/ExecutionEngine/Orc/RTDyldObjectLinkingLayer.h>
+#include <iostream>
 #include <llvm/Support/InitLLVM.h>
-#include <llvm/Support/MemoryBuffer.h>
 #include <llvm/Support/TargetSelect.h>
 #include <llvm/Support/raw_ostream.h>
+
 using namespace llvm;
-using namespace llvm::orc;
 
 namespace fs = std::filesystem;
 
-extern const char *__ry_test_current_it_name();
-
-static void test_timeout_handler(int) {
-    const char msg[] = "\nTest timed out (60s): ";
-    (void)write(STDERR_FILENO, msg, sizeof(msg) - 1);
-    const char *name = __ry_test_current_it_name();
-    if (name && name[0]) {
-        size_t len = 0;
-        while (name[len]) ++len;
-        (void)write(STDERR_FILENO, name, len);
-    }
-    const char nl[] = "\n";
-    (void)write(STDERR_FILENO, nl, 1);
-    _exit(124);
-}
-
-// RY_ENV: skip $RY_HOME/lib when running in "internal" mode
-static bool g_skip_global_lib = false;
-static bool g_trace_enabled = false;
-static std::string g_trace_out = "-";
-
-// Coverage: accumulated filenames across multiple runRyFile calls.
-// Maps global file_id to filename for user files (excluding stdlib).
-static std::vector<std::string> g_coverage_filenames;
-static int g_coverage_file_count = 0;
-// Global offset to avoid file_id collisions across separate runRyFile calls.
-static int g_coverage_file_id_offset = 0;
-
-static void resetCoverageState() {
-    __ry_coverage_reset();
-    g_coverage_file_id_offset = 0;
-    g_coverage_file_count = 0;
-    g_coverage_filenames.clear();
-}
-
-static void emitCoverageReport() {
-    if (g_coverage_file_count <= 0) return;
-    std::vector<const char*> ptrs(g_coverage_file_count);
-    for (int i = 0; i < g_coverage_file_count; ++i)
-        ptrs[i] = g_coverage_filenames[i].c_str();
-    __ry_coverage_report_summary(ptrs.data(), g_coverage_file_count);
-}
-
-// Compile and run Ry source code via JIT. Returns the JIT function's exit code.
-static int runRySource(const std::string &src, const std::string &source_name,
-                       const std::string &referrer_dir, bool test_mode,
-                       const char *argv0, bool coverage_mode = false,
-                       bool outline_mode = false) {
-    SourceLocation sourceLoc{1, 1, 0};
-    if (ry::traceEnabled()) {
-        ry::emitTraceEvent("source.loaded", "compile", &sourceLoc,
-                           {ry::TraceField("file", source_name),
-                            ry::TraceField("size_bytes", static_cast<int64_t>(src.size()))});
-    }
-
-    std::string project_root;
-    // Load .env from project root (once per root per process)
-    {
-        if (auto root = findProjectRoot(referrer_dir)) {
-            project_root = *root;
-            static std::string loaded_root;
-            if (loaded_root != *root) {
-                const char *ry_env = std::getenv("RY_ENV");
-                ry::loadDotEnv(*root, ry_env ? ry_env : "");
-                loaded_root = *root;
-            }
-        }
-    }
-
-    // Source manager for rich error messages
-    SourceManager sm;
-    int fileId = sm.addSource(source_name, src);
-    sourceLoc.file_id = fileId;
-
-    // Lex -> Parse
-    Program prog;
-    {
-        Lexer lexer(src);
-        Parser parser(lexer, &sm, fileId);
-        try {
-            if (ry::traceEnabled())
-                ry::emitTraceEvent("parse.start", "compile", &sourceLoc,
-                                   {ry::TraceField("file", source_name)});
-            prog = parser.parseProgram();
-            if (ry::traceEnabled())
-                ry::emitTraceEvent("parse.done", "compile", &sourceLoc,
-                                   {ry::TraceField("file", source_name)});
-        } catch (const DiagnosticError &e) {
-            ry::emitTraceDiagnostic("parse.error", "compile", &e.diagnostic().loc, e.diagnostic().message);
-            throw;
-        }
-    }
-
-    // Set up search paths with share/ for stdlib
-    std::vector<std::string> search_paths;
-    std::string share_dir = ry::find_share_dir(argv0, project_root, g_skip_global_lib).string();
-    if (!share_dir.empty()) {
-        search_paths.push_back(share_dir);
-        // Enable flattened imports: "from math" resolves to share/std/math/
-        search_paths.push_back((fs::path(share_dir) / "std").string());
-    }
-
-    ModuleLoader loader(search_paths, &sm);
-
-    // Implicit `from std` import (like Java's java.lang)
-    // Only insert if std can be found (graceful degradation without stdlib)
-    try {
-        Program std_prog;
-        std_prog.push_back(ImportStmt{"std", {}, {0, 0, fileId}});
-        // Resolve against stdlib search paths only (not referrer_dir)
-        // to prevent local std/ from shadowing the bundled stdlib
-        std_prog = loader.resolveImports(std_prog, "");
-        prog.insert(prog.begin(),
-            std::make_move_iterator(std_prog.begin()),
-            std::make_move_iterator(std_prog.end()));
-    } catch (const std::runtime_error &) {
-        // stdlib not installed — continue without it
-    }
-
-    // Resolve remaining imports
-    try {
-        prog = loader.resolveImports(prog, referrer_dir);
-    } catch (const DiagnosticError &) {
-        throw;
-    } catch (const std::exception &e) {
-        ry::emitTraceEvent("import.resolve.error", "compile", &sourceLoc,
-                           {ry::TraceField("file", source_name),
-                            ry::TraceField("detail", e.what())});
-        throw;
-    }
-
-    // CodeGen -> ThreadSafeModule
-    if (ry::traceEnabled())
-        ry::emitTraceEvent("codegen.start", "compile", &sourceLoc,
-                           {ry::TraceField("file", source_name)});
-    CodeGen cg(test_mode, &sm, coverage_mode, g_coverage_file_id_offset, outline_mode);
-    ThreadSafeModule tsm = [&]() {
-        try {
-            auto module = cg.compile(prog);
-            if (ry::traceEnabled())
-                ry::emitTraceEvent("codegen.done", "compile", &sourceLoc,
-                                   {ry::TraceField("file", source_name)});
-            return module;
-        } catch (const DiagnosticError &e) {
-            ry::emitTraceDiagnostic("codegen.error", "compile", &e.diagnostic().loc, e.diagnostic().message);
-            throw;
-        }
-    }();
-
-    // Build LLJIT
-    if (ry::traceEnabled())
-        ry::emitTraceEvent("jit.start", "jit", &sourceLoc,
-                           {ry::TraceField("file", source_name)});
-    auto jitOrErr = LLJITBuilder().create();
-    if (!jitOrErr) {
-        errs() << "Failed to create JIT: ";
-        logAllUnhandledErrors(jitOrErr.takeError(), errs());
-        ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                           {ry::TraceField("detail", "Failed to create JIT")});
-        return 1;
-    }
-    auto &jit = *jitOrErr;
-
-    jit->getIRTransformLayer().setTransform(ry::optimizeModule);
-
-    auto &mainJD = jit->getMainJITDylib();
-    auto dlsg = DynamicLibrarySearchGenerator::GetForCurrentProcess(
-        jit->getDataLayout().getGlobalPrefix());
-    if (!dlsg) {
-        errs() << "Failed to create DynamicLibrarySearchGenerator: ";
-        logAllUnhandledErrors(dlsg.takeError(), errs());
-        ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                           {ry::TraceField("detail", "Failed to create DynamicLibrarySearchGenerator")});
-        return 1;
-    }
-    mainJD.addGenerator(std::move(*dlsg));
-
-    // Load dynamic libraries required by @native("libname") declarations
-    {
-        // Resolve the executable path for reliable exe-relative library search.
-        // argv0 may be a bare name (e.g. "ry" from PATH), so resolve it first.
-        std::string resolvedExe = ry::self_update::detail::get_executable_path();
-        if (resolvedExe.empty()) resolvedExe = argv0;
-        const std::string &exeForLibSearch = resolvedExe;
-        auto requiredLibs = cg.getRequiredLibraries();
-        for (const auto &libName : requiredLibs) {
-            auto libPath = ry::find_native_library(exeForLibSearch, libName);
-            if (libPath.empty()) {
-                errs() << "Error: native library '" << libName
-                       << "' not found (searched exe/../lib/, exe/lib/, $RY_HOME/lib/)\n";
-                ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                                   {ry::TraceField("detail", "native library not found: " + libName)});
-                return 1;
-            }
-            if (ry::traceEnabled())
-                ry::emitTraceEvent("jit.load_library", "jit", &sourceLoc,
-                                   {ry::TraceField("library", libName),
-                                    ry::TraceField("path", libPath.string())});
-            auto libGen = DynamicLibrarySearchGenerator::Load(
-                libPath.string().c_str(),
-                jit->getDataLayout().getGlobalPrefix());
-            if (!libGen) {
-                errs() << "Failed to load native library '" << libName << "': ";
-                logAllUnhandledErrors(libGen.takeError(), errs());
-                ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                                   {ry::TraceField("detail", "Failed to load library: " + libName)});
-                return 1;
-            }
-            mainJD.addGenerator(std::move(*libGen));
-        }
-    }
-
-    // Register test runtime symbols if in test mode
-    if (test_mode) {
-        auto &es = jit->getExecutionSession();
-        SymbolMap testSymbols;
-        testSymbols[es.intern("__ry_test_describe_begin")] =
-            {ExecutorAddr::fromPtr(&__ry_test_describe_begin), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_describe_end")] =
-            {ExecutorAddr::fromPtr(&__ry_test_describe_end), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_it_begin")] =
-            {ExecutorAddr::fromPtr(&__ry_test_it_begin), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_it_end")] =
-            {ExecutorAddr::fromPtr(&__ry_test_it_end), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_expect_fail")] =
-            {ExecutorAddr::fromPtr(&__ry_test_expect_fail), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_summary")] =
-            {ExecutorAddr::fromPtr(&__ry_test_summary), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_mock_set")] =
-            {ExecutorAddr::fromPtr(&__ry_mock_set), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_mock_get")] =
-            {ExecutorAddr::fromPtr(&__ry_mock_get), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_mock_get_call_count")] =
-            {ExecutorAddr::fromPtr(&__ry_mock_get_call_count), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_mock_increment_call")] =
-            {ExecutorAddr::fromPtr(&__ry_mock_increment_call), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_mock_clear_all")] =
-            {ExecutorAddr::fromPtr(&__ry_mock_clear_all), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_prop_init_rng")] =
-            {ExecutorAddr::fromPtr(&__ry_test_prop_init_rng), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_rand_int")] =
-            {ExecutorAddr::fromPtr(&__ry_test_rand_int), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_rand_float")] =
-            {ExecutorAddr::fromPtr(&__ry_test_rand_float), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_rand_bool")] =
-            {ExecutorAddr::fromPtr(&__ry_test_rand_bool), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_rand_str")] =
-            {ExecutorAddr::fromPtr(&__ry_test_rand_str), JITSymbolFlags::Exported};
-        testSymbols[es.intern("__ry_test_it_is_failed")] =
-            {ExecutorAddr::fromPtr(&__ry_test_it_is_failed), JITSymbolFlags::Exported};
-        if (auto err = mainJD.define(absoluteSymbols(std::move(testSymbols)))) {
-            errs() << "Failed to define test symbols: ";
-            logAllUnhandledErrors(std::move(err), errs());
-            ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                               {ry::TraceField("detail", "Failed to define test symbols")});
-            return 1;
-        }
-    }
-
-    // Register coverage runtime symbols if in coverage mode
-    if (coverage_mode) {
-        auto &es = jit->getExecutionSession();
-        SymbolMap covSymbols;
-        covSymbols[es.intern("__ry_coverage_hit")] =
-            {ExecutorAddr::fromPtr(&__ry_coverage_hit), JITSymbolFlags::Exported};
-        if (auto err = mainJD.define(absoluteSymbols(std::move(covSymbols)))) {
-            errs() << "Failed to define coverage symbols: ";
-            logAllUnhandledErrors(std::move(err), errs());
-            ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                               {ry::TraceField("detail", "Failed to define coverage symbols")});
-            return 1;
-        }
-    }
-
-    // Add module
-    if (auto err = jit->addIRModule(std::move(tsm))) {
-        errs() << "Failed to add IR module: ";
-        logAllUnhandledErrors(std::move(err), errs());
-        ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                           {ry::TraceField("detail", "Failed to add IR module")});
-        return 1;
-    }
-    if (ry::traceEnabled())
-        ry::emitTraceEvent("jit.ready", "jit", &sourceLoc,
-                           {ry::TraceField("file", source_name)});
-
-    // Lookup and run __ry_main__
-    auto symOrErr = jit->lookup("__ry_main__");
-    if (!symOrErr) {
-        errs() << "Failed to lookup __ry_main__: ";
-        logAllUnhandledErrors(symOrErr.takeError(), errs());
-        ry::emitTraceEvent("runtime.error", "jit", &sourceLoc,
-                           {ry::TraceField("detail", "Failed to lookup __ry_main__")});
-        return 1;
-    }
-    auto *fn = symOrErr->toPtr<int(*)()>();
-
-    if (test_mode) {
-        std::signal(SIGALRM, test_timeout_handler);
-    }
-    if (ry::traceEnabled())
-        ry::emitTraceEvent("exec.start", "runtime", &sourceLoc,
-                           {ry::TraceField("file", source_name)});
-    int result = fn();
-    if (ry::traceEnabled())
-        ry::emitTraceEvent("exec.end", "runtime", &sourceLoc,
-                           {ry::TraceField("file", source_name),
-                            ry::TraceField("exit_code", result)});
-
-    // Record filenames for coverage reporting (accumulates across calls).
-    // Exclude stdlib files (those under share_dir) from the report.
-    if (coverage_mode) {
-        int fc = sm.getFileCount();
-        int new_total = g_coverage_file_id_offset + fc;
-        if (new_total > g_coverage_file_count) {
-            g_coverage_filenames.resize(new_total);
-            for (int i = 0; i < fc; ++i) {
-                int gid = g_coverage_file_id_offset + i;
-                const std::string &fname = sm.getFilename(i);
-                bool is_stdlib = false;
-                if (!share_dir.empty()) {
-                    auto canonical = fs::weakly_canonical(fname).string();
-                    auto canonical_share = fs::weakly_canonical(share_dir).string();
-                    is_stdlib = canonical.find(canonical_share) == 0;
-                }
-                g_coverage_filenames[gid] = is_stdlib ? "" : fname;
-            }
-            g_coverage_file_count = new_total;
-        }
-        g_coverage_file_id_offset += fc;
-    }
-
-    return result > 0 ? 1 : 0;
-}
-
-// Compile and run a .ry file via JIT. Returns the JIT function's exit code.
-static int runRyFile(const std::string &filepath, bool test_mode,
-                     const char *argv0, bool coverage_mode = false,
-                     bool outline_mode = false) {
-    auto bufOrErr = MemoryBuffer::getFile(filepath);
-    if (!bufOrErr) {
-        errs() << "Error reading file: " << filepath << "\n";
-        return 1;
-    }
-    std::string src = (*bufOrErr)->getBuffer().str();
-    std::string referrer_dir = fs::path(filepath).parent_path().string();
-    return runRySource(src, filepath, referrer_dir, test_mode, argv0, coverage_mode, outline_mode);
-}
-
-// Collect *.test.ry files recursively under root_dir
-static std::vector<std::string> findTestFiles(const std::string &root_dir) {
-    static const std::unordered_set<std::string> skip_dirs = {
-        ".git", "build", "node_modules"
-    };
-    std::vector<std::string> files;
-    std::error_code ec;
-    auto it = fs::recursive_directory_iterator(
-        root_dir, fs::directory_options::skip_permission_denied, ec);
-    if (ec) return files;
-    for (auto end = fs::recursive_directory_iterator(); it != end; it.increment(ec)) {
-        if (ec) { ec.clear(); continue; }
-        auto &entry = *it;
-        if (entry.is_directory() && skip_dirs.count(entry.path().filename().string())) {
-            it.disable_recursion_pending();
-            continue;
-        }
-        if (!entry.is_regular_file()) continue;
-        auto path = entry.path().string();
-        if (path.size() >= 8 && path.compare(path.size() - 8, 8, ".test.ry") == 0) {
-            files.push_back(path);
-        }
-    }
-    std::sort(files.begin(), files.end());
-    return files;
-}
-
-// Run multiple test files sequentially and report results
-static int runTestFilesSequential(const std::vector<std::string> &test_files,
-                                  const char *argv0, bool coverage = false,
-                                  bool outline = false) {
-    if (coverage) resetCoverageState();
-    int total_failed = 0;
-    int total_files = 0;
-    for (const auto &tf : test_files) {
-        ++total_files;
-        try {
-            int failed = runRyFile(tf, /*test_mode=*/true, argv0, coverage, outline);
-            total_failed += failed;
-        } catch (const DiagnosticError &e) {
-            errs() << e.what();
-            ++total_failed;
-        } catch (const std::exception &e) {
-            errs() << "Error in " << tf << ": " << e.what() << "\n";
-            ++total_failed;
-        }
-    }
-    if (!outline && total_files > 1) {
-        std::printf("\n%d test files executed, %d total failures\n",
-                    total_files, total_failed);
-    }
-    if (coverage) emitCoverageReport();
-    return total_failed > 0 ? 1 : 0;
-}
-
-struct TestFileResult {
-    std::string filepath;
-    int exit_code;
-    std::string output;
-};
-
-// Run a single test file in a subprocess, capturing stdout+stderr.
-// Uses posix_spawn instead of fork to be safe from multi-threaded contexts.
-static TestFileResult runTestFileSubprocess(const std::string &filepath,
-                                            const std::string &exe_path) {
-    TestFileResult result;
-    result.filepath = filepath;
-    result.exit_code = -1;
-
-    int pipefd[2];
-    if (pipe(pipefd) == -1) {
-        result.output = "Error: pipe() failed: " + std::string(strerror(errno)) + "\n";
-        return result;
-    }
-
-    posix_spawn_file_actions_t actions;
-    int rc = posix_spawn_file_actions_init(&actions);
-    if (rc == 0) rc = posix_spawn_file_actions_addclose(&actions, pipefd[0]);
-    if (rc == 0) rc = posix_spawn_file_actions_adddup2(&actions, pipefd[1], STDOUT_FILENO);
-    if (rc == 0) rc = posix_spawn_file_actions_adddup2(&actions, pipefd[1], STDERR_FILENO);
-    if (rc == 0) rc = posix_spawn_file_actions_addclose(&actions, pipefd[1]);
-    if (rc != 0) {
-        posix_spawn_file_actions_destroy(&actions);
-        close(pipefd[0]);
-        close(pipefd[1]);
-        result.output = "Error: posix_spawn_file_actions failed: " + std::string(strerror(rc)) + "\n";
-        return result;
-    }
-
-    const char *argv[] = {exe_path.c_str(), "test", filepath.c_str(), nullptr};
-    pid_t pid;
-    int err = posix_spawn(&pid, exe_path.c_str(), &actions, nullptr,
-                          const_cast<char *const *>(argv), RY_ENVIRON);
-    posix_spawn_file_actions_destroy(&actions);
-
-    if (err != 0) {
-        close(pipefd[0]);
-        close(pipefd[1]);
-        result.output = "Error: posix_spawn() failed: " + std::string(strerror(err)) + "\n";
-        return result;
-    }
-
-    close(pipefd[1]);
-    char buf[4096];
-    ssize_t n;
-    while ((n = read(pipefd[0], buf, sizeof(buf))) > 0 || (n == -1 && errno == EINTR)) {
-        if (n > 0) result.output.append(buf, n);
-    }
-    close(pipefd[0]);
-
-    int status = 0;
-    pid_t wp;
-    while ((wp = waitpid(pid, &status, 0)) == -1) {
-        if (errno != EINTR) break;
-    }
-
-    if (wp == -1)
-        result.exit_code = -1;
-    else if (WIFEXITED(status))
-        result.exit_code = WEXITSTATUS(status);
-    else if (WIFSIGNALED(status))
-        result.exit_code = 128 + WTERMSIG(status);
-    else
-        result.exit_code = -1;
-
-    return result;
-}
-
-// Run multiple test files in parallel using worker threads
-static int runTestFilesParallel(const std::vector<std::string> &test_files,
-                                const std::string &exe_path, int parallelism) {
-    int num_files = static_cast<int>(test_files.size());
-    std::vector<TestFileResult> results(num_files);
-
-    // Work queue: indices into test_files
-    std::deque<int> work_queue;
-    for (int i = 0; i < num_files; ++i)
-        work_queue.push_back(i);
-
-    std::mutex queue_mutex;
-    std::mutex progress_mutex;
-    int completed = 0;
-
-    auto worker = [&]() {
-        while (true) {
-            int idx;
-            {
-                std::lock_guard<std::mutex> lock(queue_mutex);
-                if (work_queue.empty()) return;
-                idx = work_queue.front();
-                work_queue.pop_front();
-            }
-            results[idx] = runTestFileSubprocess(test_files[idx], exe_path);
-            {
-                std::lock_guard<std::mutex> lock(progress_mutex);
-                ++completed;
-                std::fprintf(stderr, "\r\033[K[%d/%d] Running tests...",
-                             completed, num_files);
-            }
-        }
-    };
-
-    int num_workers = std::min(parallelism, num_files);
-    std::vector<std::thread> threads;
-    threads.reserve(num_workers);
-
-    auto wall_start = std::chrono::steady_clock::now();
-    for (int i = 0; i < num_workers; ++i)
-        threads.emplace_back(worker);
-    for (auto &t : threads)
-        t.join();
-    auto wall_end = std::chrono::steady_clock::now();
-    double wall_elapsed = std::chrono::duration<double>(wall_end - wall_start).count();
-
-    std::fprintf(stderr, "\r\033[K");
-
-    int total_failed = 0;
-    for (const auto &r : results) {
-        std::fputs(r.output.c_str(), stdout);
-        if (r.exit_code != 0)
-            ++total_failed;
-    }
-
-    std::printf("\n%d test files executed, %d total failures (%.2fs, %d workers)\n",
-                num_files, total_failed, wall_elapsed, num_workers);
-    return total_failed > 0 ? 1 : 0;
-}
-
-// Run multiple test files and report results
-static int runTestFiles(const std::vector<std::string> &test_files,
-                        const char *argv0, bool parallel, bool coverage = false,
-                        bool outline = false) {
-    if (outline || !parallel || test_files.size() <= 1) {
-        return runTestFilesSequential(test_files, argv0, coverage, outline);
-    }
-    std::string exe_path = ry::self_update::detail::get_executable_path();
-    if (exe_path.empty()) {
-        errs() << "Warning: cannot resolve executable path, falling back to sequential\n";
-        return runTestFilesSequential(test_files, argv0, coverage);
-    }
-    unsigned hw = std::thread::hardware_concurrency();
-    int parallelism = static_cast<int>(std::min({hw, 8u, static_cast<unsigned>(test_files.size())}));
-    if (parallelism < 1) parallelism = 1;
-    return runTestFilesParallel(test_files, exe_path, parallelism);
-}
-
-// Discover *.test.ry files in a directory and run them
-static int discoverAndRunTests(const std::string &dir, const char *argv0,
-                               bool parallel, bool coverage = false,
-                               bool outline = false) {
-    auto test_files = findTestFiles(dir);
-    if (test_files.empty()) {
-        errs() << "No *.test.ry files found in " << dir << "\n";
-        return 1;
-    }
-    return runTestFiles(test_files, argv0, parallel, coverage, outline);
-}
-
-static void applyRyEnv(const char *value) {
-    std::string canonical = ry::resolveRyEnv(value);
-    setenv("RY_ENV", canonical.c_str(), 1);
-    if (canonical == "internal")
-        g_skip_global_lib = true;
-}
-
-static bool isHelpFlag(const char *arg) {
-    return std::strcmp(arg, "-h") == 0 || std::strcmp(arg, "--help") == 0;
-}
-
-static bool hasHelpFlag(int argc, char *argv[], int start = 1) {
-    for (int i = start; i < argc; ++i)
-        if (isHelpFlag(argv[i])) return true;
-    return false;
-}
-
-static bool isKnownSubcommand(const char *arg) {
-    static const char *known[] = {
-        "test", "init", "new", "fmt", "self-update", "run", nullptr
-    };
-    for (const char **p = known; *p; ++p)
-        if (std::strcmp(arg, *p) == 0) return true;
-    return false;
-}
-
-// Resolve the entry point file path from package.toml.
-// Returns the absolute path on success, or empty string if not found.
-// When require is true, error messages are written to stderr.
-static std::string resolveEntryPoint(bool require) {
-    auto root = findProjectRoot();
-    if (!root) {
-        if (require) errs() << "Error: package.toml not found. Run 'ry init' first.\n";
-        return "";
-    }
-    std::ifstream f(fs::path(*root) / "package.toml");
-    if (!f) {
-        if (require) errs() << "Error: cannot open package.toml\n";
-        return "";
-    }
-    std::string content((std::istreambuf_iterator<char>(f)),
-                         std::istreambuf_iterator<char>{});
-    ProjectConfig config;
-    try {
-        config = ProjectConfigParser::load(content);
-    } catch (const std::exception &e) {
-        if (require) errs() << "Error: failed to parse package.toml: " << e.what() << "\n";
-        return "";
-    }
-    if (config.entry.empty()) {
-        if (require) errs() << "Error: no 'entry' field in package.toml\n";
-        return "";
-    }
-    std::string path = (fs::path(*root) / config.entry).string();
-    if (!fs::exists(path)) {
-        if (require) errs() << "Error: entry file not found: " << config.entry << "\n";
-        return "";
-    }
-    return path;
-}
-
-static void printMainHelp() {
-    llvm::outs() << "ry " << RY_VERSION << "\n\n";
-    llvm::outs() << "Usage:\n";
-    llvm::outs() << "  ry                                   Run entry point from package.toml\n";
-    llvm::outs() << "  ry <file.ry> [args...]               Run a Ry script\n";
-    llvm::outs() << "  echo '<code>' | ry -c                Run code from stdin\n";
-    llvm::outs() << "  ry test [options] [<file> | <dir>]   Run tests\n";
-    llvm::outs() << "  ry init                              Initialize a project\n";
-    llvm::outs() << "  ry new <project-name>                Create a new project\n";
-    llvm::outs() << "  ry run [<script-name>]               Run a project script\n";
-    llvm::outs() << "  ry fmt [options] [<file> | <dir>]    Format source files\n";
-    llvm::outs() << "  ry self-update [options]              Update ry itself\n";
-    llvm::outs() << "\n";
-    llvm::outs() << "Options:\n";
-    llvm::outs() << "  -c               Read and execute code from stdin\n";
-    llvm::outs() << "  -h, --help       Show this help\n";
-    llvm::outs() << "  -v, --version    Show version\n";
-    llvm::outs() << "  --env=<env>      Set environment (production|development|internal)\n";
-    llvm::outs() << "  --trace          Emit structured internal trace as JSON Lines\n";
-    llvm::outs() << "  --trace-out=PATH Write trace output to PATH (or '-' for stderr)\n";
-    llvm::outs() << "\n";
-    llvm::outs() << "Run 'ry <command> --help' for more information on a command.\n";
-}
-
-static void printTestHelp() {
-    llvm::outs() << "Usage: ry test [options] [<file.ry> | <dir>]\n\n";
-    llvm::outs() << "Run test files (*.test.ry).\n\n";
-    llvm::outs() << "Options:\n";
-    llvm::outs() << "  -p, --parallel       Run tests in parallel\n";
-    llvm::outs() << "  -w, --watch          Watch for changes and re-run\n";
-    llvm::outs() << "  --coverage, --cov    Collect coverage information\n";
-    llvm::outs() << "  --outline            Print describe/it structure without running tests\n";
-    llvm::outs() << "  --trace              Emit structured internal trace as JSON Lines\n";
-    llvm::outs() << "  --trace-out=PATH     Write trace output to PATH (or '-' for stderr)\n";
-    llvm::outs() << "  -h, --help           Show this help\n";
-}
-
-static void printInitHelp() {
-    llvm::outs() << "Usage: ry init\n\n";
-    llvm::outs() << "Initialize a new Ry project in the current directory.\n";
-    llvm::outs() << "Creates a package.toml and basic project structure.\n";
-}
-
-static void printNewHelp() {
-    llvm::outs() << "Usage: ry new <project-name>\n\n";
-    llvm::outs() << "Create a new Ry project in a new directory.\n";
-}
-
-static void printFmtHelp() {
-    llvm::outs() << "Usage: ry fmt [options] [<file> | <dir>]\n\n";
-    llvm::outs() << "Format Ry source files.\n\n";
-    llvm::outs() << "Options:\n";
-    llvm::outs() << "  --check      Check formatting without modifying files\n";
-    llvm::outs() << "  -h, --help   Show this help\n";
-}
-
-static void printRunHelp() {
-    llvm::outs() << "Usage: ry run [<script-name>]\n\n";
-    llvm::outs() << "Run a script defined in package.toml [scripts] section.\n\n";
-    llvm::outs() << "If no script name is given, lists all available scripts.\n";
-}
-
-static void printSelfUpdateHelp() {
-    llvm::outs() << "Usage: ry self-update [--nightly | <version>]\n\n";
-    llvm::outs() << "Update ry to a newer version.\n\n";
-    llvm::outs() << "Options:\n";
-    llvm::outs() << "  (no args)    Update to latest stable release\n";
-    llvm::outs() << "  --nightly    Update to latest nightly/prerelease\n";
-    llvm::outs() << "  <version>    Update to a specific version (e.g. v0.0.1)\n";
-    llvm::outs() << "  -h, --help   Show this help\n";
-}
-
-static void parseRyEnv(int &argc, char **&argv) {
-    // --env=<value> CLI flag takes precedence over RY_ENV env var
-    if (argc >= 2 && std::strncmp(argv[1], "--env=", 6) == 0) {
-        applyRyEnv(argv[1] + 6);
-        // Remove --env flag from argv, preserving argv[0] (program name)
-        argv[1] = argv[0];
-        ++argv;
-        --argc;
-    } else if (const char *env = std::getenv("RY_ENV")) {
-        applyRyEnv(env);
-    }
-}
-
-static void parseGlobalFlags(int &argc, char **&argv) {
-    while (argc >= 2) {
-        if (std::strcmp(argv[1], "--trace") == 0) {
-            g_trace_enabled = true;
-        } else if (std::strncmp(argv[1], "--trace-out=", 12) == 0) {
-            g_trace_enabled = true;
-            g_trace_out = argv[1] + 12;
-        } else {
-            break;
-        }
-
-        argv[1] = argv[0];
-        ++argv;
-        --argc;
-    }
-}
-
 int main(int argc, char *argv[]) {
-    parseRyEnv(argc, argv);
-    parseGlobalFlags(argc, argv);
-    ry::configureTrace(g_trace_enabled, g_trace_out);
+    bool skip_global_lib = ry::cli::parseRyEnv(argc, argv);
+    bool trace_enabled = false;
+    std::string trace_out = "-";
+    ry::cli::parseGlobalFlags(argc, argv, trace_enabled, trace_out);
+    ry::configureTrace(trace_enabled, trace_out);
     bool sessionStarted = false;
     if (ry::traceEnabled()) {
         ry::emitTraceEvent("session.start", "session", nullptr,
@@ -804,20 +43,20 @@ int main(int argc, char *argv[]) {
 
     // Help flag handling: -h/--help takes priority over other options
     if (argc >= 2) {
-        if (isKnownSubcommand(argv[1])) {
+        if (ry::cli::isKnownSubcommand(argv[1])) {
             // Subcommand help: ry <subcmd> ... -h/--help
-            if (hasHelpFlag(argc, argv, 2)) {
-                if (std::strcmp(argv[1], "test") == 0) { printTestHelp(); return 0; }
-                if (std::strcmp(argv[1], "init") == 0) { printInitHelp(); return 0; }
-                if (std::strcmp(argv[1], "new") == 0) { printNewHelp(); return 0; }
-                if (std::strcmp(argv[1], "fmt") == 0) { printFmtHelp(); return 0; }
-                if (std::strcmp(argv[1], "run") == 0) { printRunHelp(); return 0; }
-                if (std::strcmp(argv[1], "self-update") == 0) { printSelfUpdateHelp(); return 0; }
+            if (ry::cli::hasHelpFlag(argc, argv, 2)) {
+                if (std::strcmp(argv[1], "test") == 0) { ry::cli::printTestHelp(); return 0; }
+                if (std::strcmp(argv[1], "init") == 0) { ry::cli::printInitHelp(); return 0; }
+                if (std::strcmp(argv[1], "new") == 0) { ry::cli::printNewHelp(); return 0; }
+                if (std::strcmp(argv[1], "fmt") == 0) { ry::cli::printFmtHelp(); return 0; }
+                if (std::strcmp(argv[1], "run") == 0) { ry::cli::printRunHelp(); return 0; }
+                if (std::strcmp(argv[1], "self-update") == 0) { ry::cli::printSelfUpdateHelp(); return 0; }
             }
         } else {
             // Main help: ry -h, ry --help, ry --env=internal -h, etc.
-            if (hasHelpFlag(argc, argv)) {
-                printMainHelp();
+            if (ry::cli::hasHelpFlag(argc, argv)) {
+                ry::cli::printMainHelp();
                 return 0;
             }
         }
@@ -863,7 +102,7 @@ int main(int argc, char *argv[]) {
         __ry_args_init(0, nullptr);
         std::string cwd = fs::current_path().string();
         try {
-            return runRySource(src, "<stdin>", cwd, false, argv[0]);
+            return ry::runRySource(src, "<stdin>", cwd, false, argv[0], skip_global_lib);
         } catch (const DiagnosticError &e) {
             errs() << e.what();
             return 1;
@@ -873,7 +112,7 @@ int main(int argc, char *argv[]) {
         }
     } else if (argc >= 2 && std::strcmp(argv[1], "--") == 0) {
         // ry -- [args...] — run entry point with arguments
-        entry_path_storage = resolveEntryPoint(true);
+        entry_path_storage = ry::cli::resolveEntryPoint(true);
         if (entry_path_storage.empty()) return 1;
         filename = entry_path_storage.c_str();
         __ry_args_init(argc - 2, argc > 2 ? argv + 2 : nullptr);
@@ -883,7 +122,7 @@ int main(int argc, char *argv[]) {
         std::string arg1 = argv[1];
         if (!fs::exists(arg1)) {
             errs() << "Error: unknown command '" << arg1 << "'\n\n";
-            printMainHelp();
+            ry::cli::printMainHelp();
             return 1;
         }
         filename = argv[1];
@@ -894,7 +133,7 @@ int main(int argc, char *argv[]) {
         bool watch = false;
         bool coverage = false;
         bool outline = false;
-        std::string trace_out = g_trace_out;
+        std::string test_trace_out = trace_out;
         const char *target = nullptr;
         for (int i = 2; i < argc; ++i) {
             if (std::strcmp(argv[i], "-p") == 0 ||
@@ -909,16 +148,16 @@ int main(int argc, char *argv[]) {
             } else if (std::strcmp(argv[i], "--outline") == 0) {
                 outline = true;
             } else if (std::strcmp(argv[i], "--trace") == 0) {
-                g_trace_enabled = true;
+                trace_enabled = true;
             } else if (std::strncmp(argv[i], "--trace-out=", 12) == 0) {
-                g_trace_enabled = true;
-                trace_out = argv[i] + 12;
+                trace_enabled = true;
+                test_trace_out = argv[i] + 12;
             } else if (!target) {
                 target = argv[i];
             }
         }
-        if (g_trace_enabled) {
-            ry::configureTrace(true, trace_out);
+        if (trace_enabled) {
+            ry::configureTrace(true, test_trace_out);
             if (ry::traceEnabled() && !sessionStarted) {
                 ry::emitTraceEvent("session.start", "session", nullptr,
                                    {ry::TraceField("argv0", argc > 0 && argv[0] ? argv[0] : "ry")});
@@ -929,7 +168,7 @@ int main(int argc, char *argv[]) {
             errs() << "Warning: --coverage is not supported with --parallel, falling back to sequential\n";
             parallel = false;
         }
-        if (g_trace_enabled && parallel) {
+        if (trace_enabled && parallel) {
             errs() << "Warning: --trace is not supported with --parallel, falling back to sequential\n";
             parallel = false;
         }
@@ -940,12 +179,13 @@ int main(int argc, char *argv[]) {
             if (fs::is_directory(target_str, ec)) {
                 if (watch) {
                     const char *a0 = argv[0];
-                    ry::watchAndRunTests(target_str, [target_str, a0, parallel, coverage, outline]() {
-                        discoverAndRunTests(target_str, a0, parallel, coverage, outline);
+                    bool sgl = skip_global_lib;
+                    ry::watchAndRunTests(target_str, [target_str, a0, sgl, parallel, coverage, outline]() {
+                        ry::discoverAndRunTests(target_str, a0, sgl, parallel, coverage, outline);
                     });
                     return 0;
                 }
-                return discoverAndRunTests(target_str, argv[0], parallel, coverage, outline);
+                return ry::discoverAndRunTests(target_str, argv[0], skip_global_lib, parallel, coverage, outline);
             }
             if (ec) {
                 errs() << "Error: cannot access " << target_str << ": "
@@ -957,9 +197,10 @@ int main(int argc, char *argv[]) {
                 auto target_dir = fs::path(target_str).parent_path().string();
                 std::string watch_root = findProjectRoot(target_dir).value_or(target_dir);
                 const char *a0 = argv[0];
-                ry::watchAndRunTests(watch_root, [target_str, a0, outline]() {
+                bool sgl = skip_global_lib;
+                ry::watchAndRunTests(watch_root, [target_str, a0, sgl, outline]() {
                     try {
-                        runRyFile(target_str, true, a0, false, outline);
+                        ry::runRyFile(target_str, true, a0, sgl, false, outline);
                     } catch (const DiagnosticError &e) {
                         errs() << e.what();
                     } catch (const std::exception &e) {
@@ -984,17 +225,18 @@ int main(int argc, char *argv[]) {
             if (watch) {
                 std::string root_dir = *root;
                 const char *a0 = argv[0];
-                ry::watchAndRunTests(root_dir, [root_dir, a0, parallel, coverage, outline]() {
-                    discoverAndRunTests(root_dir, a0, parallel, coverage, outline);
+                bool sgl = skip_global_lib;
+                ry::watchAndRunTests(root_dir, [root_dir, a0, sgl, parallel, coverage, outline]() {
+                    ry::discoverAndRunTests(root_dir, a0, sgl, parallel, coverage, outline);
                 });
                 return 0;
             }
-            return discoverAndRunTests(*root, argv[0], parallel, coverage, outline);
+            return ry::discoverAndRunTests(*root, argv[0], skip_global_lib, parallel, coverage, outline);
         }
     } else if (argc == 1) {
-        entry_path_storage = resolveEntryPoint(false);
+        entry_path_storage = ry::cli::resolveEntryPoint(false);
         if (entry_path_storage.empty()) {
-            printMainHelp();
+            ry::cli::printMainHelp();
             return 1;
         }
         filename = entry_path_storage.c_str();
@@ -1002,9 +244,11 @@ int main(int argc, char *argv[]) {
     }
 
     try {
-        if (coverage_mode) resetCoverageState();
-        int rc = runRyFile(filename, test_mode, argv[0], coverage_mode, outline_mode);
-        if (coverage_mode) emitCoverageReport();
+        ry::CoverageState cs;
+        if (coverage_mode) ry::resetCoverageState(cs);
+        int rc = ry::runRyFile(filename, test_mode, argv[0], skip_global_lib,
+                               coverage_mode, outline_mode, coverage_mode ? &cs : nullptr);
+        if (coverage_mode) ry::emitCoverageReport(cs);
         return rc;
     } catch (const DiagnosticError &e) {
         errs() << e.what();
